requirements.txt
translate
pyyaml
openai-whisper
ffmpeg-python
